boss enemy implemented

+++ src/core/ecs/entity.h
@@ -3,7 +3,7 @@
 
 struct EntityManager {
     // Tracks which components each entity has
-    uint32_t componentMasks[MAX_ENTITIES];
+    uint64_t componentMasks[MAX_ENTITIES];
     // Tracks which entities are active
     bool activeEntities[MAX_ENTITIES];
     // Number of active entities
